Make remove also have a 10 user limit

+++ commands/remove.js
@@ -9,21 +9,29 @@ module.exports = {
     }
     const taggedUsers = message.mentions.users;
 
-    taggedUsers.forEach(async function(user) {
-      try {
-        const response = await request.delete(
-          `${process.env.API}api/users/${user.id}.json`,
-          {
-            auth: {
-              bearer: process.env.elroy
-            },
-            form: {}
-          }
-        );
-        message.channel.send(`Arf! I removed <@${user.id}> from the list.`);
-      } catch (error) {
-        message.channel.send(`Arf! I couldn't find <@${user.id}> on the list.`);
-      }
-    });
+    if (taggedUsers.array.length <= 10) {
+      taggedUsers.forEach(async function(user) {
+        try {
+          const response = await request.delete(
+            `${process.env.API}api/users/${user.id}.json`,
+            {
+              auth: {
+                bearer: process.env.elroy
+              },
+              form: {}
+            }
+          );
+          message.channel.send(`Arf! I removed <@${user.id}> from the list.`);
+        } catch (error) {
+          message.channel.send(
+            `Arf! I couldn't find <@${user.id}> on the list.`
+          );
+        }
+      });
+    } else {
+      message.reply(
+        "I can only remove 10 people at a time. Try to tag fewer people!"
+      );
+    }
   }
 };
